+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/recipe/VanillaRecipeCache.java
@@ -4,6 +4,7 @@ import com.codetaylor.mc.artisanworktables.api.internal.recipe.*;
 import com.codetaylor.mc.artisanworktables.api.internal.reference.EnumTier;
 import com.codetaylor.mc.artisanworktables.api.recipe.ArtisanRecipe;
 import com.codetaylor.mc.artisanworktables.api.recipe.IArtisanRecipe;
+import com.codetaylor.mc.artisanworktables.modules.worktables.ModuleWorktablesConfig;
 import com.codetaylor.mc.artisanworktables.modules.worktables.integration.crafttweaker.CTArtisanIngredient;
 import com.codetaylor.mc.artisanworktables.modules.worktables.integration.crafttweaker.CTArtisanRecipe;
 import com.codetaylor.mc.artisanworktables.modules.worktables.tile.spi.TileEntityBase;
@@ -67,7 +68,7 @@ public class VanillaRecipeCache {
       recipe = CraftingManager.findMatchingRecipe(inventoryWrapper, world);
     }
 
-    if (recipe == null) {
+    if (recipe == null || ModuleWorktablesConfig.isRecipeBlacklisted(recipe.getRegistryName())) {
       return null;
     }